pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "clitutor"
version = "1.0.0"
description = "An interactive CLI tutorial with a gamified TUI"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
dependencies = [
    "textual>=0.47.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
clitutor = "clitutor.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
clitutor = ["lessons/*.md", "lessons/*.json", "assets/*", "styles/*.tcss"]
